Add map_primitive cpp file
